# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Eleventh Press International Publishing — Public API",
    "version": "1.0.0",
    "description": "The public, external-facing surface of the platform: article discovery/browsing, OAI-PMH 2.0 harvesting, COUNTER 5 / SUSHI usage reporting, and OJS-compatible export. This covers only endpoints meant for third-party/programmatic consumption — the dashboard's internal editorial-workflow API is intentionally out of scope. There is currently one implicit, unversioned API surface (no /v1/ prefix); a breaking change to any endpoint documented here would be introduced as a new path rather than silently changing this one, until a real versioning scheme is adopted.",
    "contact": { "email": "editorial@eleventhpress.org" }
  },
  "servers": [{ "url": "https://eleventhpress.vercel.app", "description": "Production (override via APP_BASE_URL if self-hosted on a custom domain)" }],
...
  "paths": {
...
    "/api/oai-pmh": {
      "get": {
        "tags": ["OAI-PMH"],
        "summary": "OAI-PMH 2.0 endpoint",
        "description": "Implements Identify, ListMetadataFormats, ListRecords, ListIdentifiers, GetRecord, and ListSets over Dublin Core (oai_dc) records for every PUBLISHED article. Selective harvesting via from/until; pagination via resumptionToken.",
        "parameters": [
          { "name": "verb", "in": "query", "required": true, "schema": { "type": "string", "enum": ["Identify", "ListMetadataFormats", "ListRecords", "ListIdentifiers", "GetRecord", "ListSets"] } },
          { "name": "metadataPrefix", "in": "query", "schema": { "type": "string", "enum": ["oai_dc"] }, "description": "Required for ListRecords/ListIdentifiers/GetRecord" },
          { "name": "identifier", "in": "query", "schema": { "type": "string" }, "description": "Required for GetRecord — the oai:<host>:<articleId> identifier from a record's <header>" },
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "until", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "resumptionToken", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "OAI-PMH XML response", "content": { "application/xml": {} } },
          "400": { "description": "OAI-PMH error response (badVerb, badArgument, badResumptionToken, idDoesNotExist, noRecordsMatch)" }
        }
      }
    },
...
  }
}